+++ src/main/timestr.c
@@ -89,7 +89,7 @@ static int hour_fill(char *bitmap, const char *tm)
 	end = -1;
 	if ((p = strchr(tm, '-')) != NULL) {
 		p++;
-		if (p - tm != 5 || strlen(p) < 4 || !isdigit(*p))
+		if (p - tm != 5 || strlen(p) < 4 || !isdigit((int) *p))
 			return 0;
 		end = 600 * val(p[0]) + 60 * val(p[1]) + atoi(p + 2);
 	}
@@ -97,7 +97,7 @@ static int hour_fill(char *bitmap, const char *tm)
 		start = 0;
 		end = DAYMIN - 1;
 	} else {
-		if (strlen(tm) < 4 || !isdigit(*tm))
+		if (strlen(tm) < 4 || !isdigit((int) *tm))
 			return 0;
 		start = 600 * val(tm[0]) + 60 * val(tm[1]) + atoi(tm + 2);
 		if (end < 0) end = start;
@@ -136,7 +136,7 @@ static int day_fill(char *bitmap, const char *tm)
 	int start, end;
 
 	for (hr = tm; *hr; hr++)
-		if (isdigit(*hr))
+		if (isdigit((int) *hr))
 			break;
 	if (hr == tm) 
 		tm = "Al";
